Working on anonymous replies

+++ modules/talk/src/Talk.php
@@ -109,10 +109,15 @@ class Talk
             $receipt->delete();
         });
 
-        $user->replies()->each(function ($reply) {
-            $reply->delete();
-            #TODO: Delete children replies
-        });
+        if ($user->anonymous_replies) {
+            $user->replies()->each(function ($reply) {
+                
+            });
+        } else {
+            $user->replies()->each(function ($reply) {
+                $reply->delete();
+            });            
+        }
 
         $user->participated_conversations()->detach();
